requirements.txt
flask==2.2.5
flask-restx==1.0.6
flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
pytest==7.3.1
